SVN source-code repository check-in for fix to VA State lines 10 and 11, switch for Std. versus Itemized Deductions.

+++ OTS_2020/src/taxsolve_f8606.c
@@ -139,7 +139,7 @@ int main(int argc, char *argv[])
       double fractional = modf(L[10], &integral);
       int integral_int, fractional_int;
       integral_int = integral;
-      fractional_int = abs(1000.0 * fractional);
+      fractional_int = abs( (int)(1000.0 * fractional) );
       fprintf(outfile, "L10intpart %d\n", integral_int);
       fprintf(outfile, "L10rest %03d\n", fractional_int);
